% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine.debiased.hetero}
\alias{combine.debiased.hetero}
\title{combine by debiasing, heterochronous time}
\usage{
combine.debiased.hetero(
  logpath,
  treepath,
  time.offset = NULL,
  tree.prior = "skyline",
  num.samples = 1000,
  recon.len = 100,
  skip = 3
)
}
\arguments{
\item{logpath}{a list of paths of the log files}

\item{treepath}{a list of paths of the tree files}

\item{tree.prior}{c("skyline", "skyride") #TODO: add skygrid}

\item{num.samples}{number of samples (from the last sample) to use}

\item{recon.len}{number of grid points for the reconstructed population size}

\item{skip}{number of rows to skip when reading the log files}

\item{M}{number of groups in skyline}

\item{fct}{a scaling factor, the population sizes and times are multiplied by fct #TODO: add fct}
}
\value{
time points and the corresponding population sizes
}
\description{
combine by debiasing, heterochronous time
}
