% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine.debiased}
\alias{combine.debiased}
\title{test}
\usage{
combine.debiased(
  logpath,
  treepath,
  skip = 3,
  tree.prior = "skyline",
  recon.len = 100,
  M = 15
)
}
\arguments{
\item{logpath}{a list of paths of the log files}

\item{treepath}{a list of paths of the tree files}

\item{skip}{number of rows to skip when reading the log files}

\item{tree.prior}{c("skyline", "skyride") #TODO: add skygrid}

\item{recon.len}{number of grid points for the reconstructed population size}

\item{M}{number of groups in skyline}

\item{fct}{a scaling factor, the population sizes and times are multiplied by fct #TODO: add fct}

\item{num.samples}{number of samples (from the last sample) to use}
}
\value{
time points and the corresponding population sizes
}
\description{
test
}
