% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{combine}
\alias{combine}
\title{Wrapper of different combining algorithms}
\usage{
combine(
  logpath,
  treepath,
  algo = "debias",
  time.offset = NULL,
  num.samples = 1000,
  recon.len = 100,
  skip = 3,
  tree.prior = "skyline",
  maxiter = 1000,
  fct = 1
)
}
\arguments{
\item{logpath}{a list of names of files containing BEAST log outputs}

\item{treepath}{a list of names of files containing BEAST trees outputs}

\item{algo}{c("debias", "gaussian", "weiszfeld")}

\item{time.offset}{a vector of length K, to be added to the times of each subsets}

\item{num.samples}{only the last num.samples samples are used}

\item{recon.len}{number of grids for the reconstructed population size}

\item{skip}{number of lines to skip when reading the BEAST log outputs}

\item{tree.prior}{c("skyline", "skyride")}

\item{maxiter}{maximum number of iterations of Weiszfeld algorithm}

\item{fct}{a factor to be multiplied to the times and population sizes}
}
\value{
reconstructed population size
}
\description{
Wrapper of different combining algorithms
}
