% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{Gaussian.approx}
\alias{Gaussian.approx}
\title{Gaussian approximation}
\usage{
Gaussian.approx(f.all)
}
\arguments{
\item{f.all}{an array with shape (K, recon.len, num.samples), where K is the number of subsets, recon.len is the number of grid points, num.samples is the number of samples used}
}
\value{
reconstructed population size
}
\description{
Gaussian approximation
}
