% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{N.standardize}
\alias{N.standardize}
\title{Project a vector of population size to a specified grid}
\usage{
N.standardize(t.i, N.i, time.grid)
}
\arguments{
\item{t.i}{vector of time points at which the population sizes are N.i}

\item{N.i}{vector of population sizes at time t.i}

\item{time.grid}{grid of time}
}
\value{
a vector of population sizes at time.grid
}
\description{
Project a vector of population size to a specified grid
}
