% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process.beast.logs}
\alias{process.beast.logs}
\title{test}
\usage{
process.beast.logs(
  logpath,
  treepath = NULL,
  M = 15,
  fct = 1,
  skip = 3,
  tree.prior = "skyline",
  num.samples = 1000,
  recon.len = 30,
  coal.return = FALSE
)
}
\arguments{
\item{logpath}{a list of paths of the log files}

\item{treepath}{a list of paths of the tree files}

\item{M}{number of groups in skyline}

\item{fct}{a scaling factor, the population sizes and times are multiplied by fct}

\item{skip}{number of rows to skip when reading the log files}

\item{tree.prior}{c("skyline", "skyride") #TODO: add skygrid}

\item{num.samples}{number of samples (from the last sample) to use}

\item{recon.len}{number of grid points for the reconstructed population size}

\item{coal.return}{return coalescent times or not}
}
\value{
f.all
}
\description{
test
}
