% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{Weiszfeld}
\alias{Weiszfeld}
\title{Weiszfeld algorithm}
\usage{
Weiszfeld(f.all, maxiter = 1000)
}
\arguments{
\item{f.all}{an array with shape (K, recon.len, num.samples), where K is the number of subsets, recon.len is the number of grid points, num.samples is the number of samples used}

\item{maxiter}{maximum number of iterations for Weiszfeld algorithm}
}
\value{
reconstructed population size
}
\description{
Weiszfeld algorithm
}
\examples{
test_fun(1)
}
