% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine.debiased.hetero}
\alias{combine.debiased.hetero}
\title{combine by the debiasing procedure}
\usage{
combine.debiased.hetero(
  logpath,
  treepath,
  fct = 1,
  time.offset = NULL,
  tree.prior = "skyline",
  num.samples = 1000,
  recon.len = 100,
  skip = 3
)
}
\arguments{
\item{logpath}{a list of paths of the log files}

\item{treepath}{a list of paths of the tree files}

\item{fct}{a scaling factor, the population sizes and times are multiplied by fct #TODO: add fct}

\item{time.offset}{a vector of length K, to be added to the times of each subsets}

\item{tree.prior}{c("skyline", "skyride") #TODO: add skygrid}

\item{num.samples}{number of samples (from the last sample) to use}

\item{recon.len}{number of grid points for the reconstructed population size}

\item{skip}{number of rows to skip when reading the log files}
}
\value{
a time grid of length recon.len and the reconstructed population sizes at corresponding time grids
}
\description{
combine by the debiasing procedure
}
