% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process.beast.logs}
\alias{process.beast.logs}
\title{preprocess the output files of BEAST}
\usage{
process.beast.logs(
  logpath,
  treepath = NULL,
  fct = 1,
  time.offset = NULL,
  tree.prior = "skyline",
  num.samples = 1000,
  recon.len = 100,
  skip = 3,
  coal.return = FALSE
)
}
\arguments{
\item{logpath}{a list of paths of the log files}

\item{treepath}{a list of paths of the tree files}

\item{fct}{a scaling factor, the population sizes and times are multiplied by fct}

\item{time.offset}{a vector of length K, to be added to the times of each subsets}

\item{tree.prior}{c("skyline", "skyride") #TODO: add skygrid}

\item{num.samples}{number of samples (from the last sample) to use}

\item{recon.len}{number of grid points for the reconstructed population size}

\item{skip}{number of rows to skip when reading the log files}

\item{coal.return}{return coalescent times or not}
}
\value{
f.all
}
\description{
preprocess the output files of BEAST
}
